requirements.txt
# OpenNemesis Dependencies

# Core
python-dotenv>=0
nest-asyncio>=1.5.0

# Telegram Bot
python-telegram-bot>=20.0

# Google Gemini AI
google-genai>=0.3.0

# Groq Whisper (transcripción)
groq>=1.0.0

# Web Search
googlesearch-python>=1.1.0

# Weather
geopy>=2.4.0
requests>=2.31.0

# Audio processing
pydub>=0.25.1
edge-tts>=7.0.0

# HTTP client
httpx>=0.25.0
